Use enum as a value, generate type for types
