+++ package.json
@@ -37,7 +37,7 @@
     "typescript": "^5.5.3"
   },
   "dependencies": {
-    "@ingenierias-lentas/recopiladora-types": "^0.0.1",
+    "@ingenierias-lentas/recopiladora-types": "^0.0.3",
     "@koa/router": "^12.0.1",
     "dotenv": "^16.4.5",
     "graphql-http": "^1.22.1",
