+++ packages/types/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@ingenierias-lentas/recopiladora-types",
-  "version": "0.0.1",
+  "version": "0.0.3",
   "description": "typescript types for recopiladora",
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
