+++ packages/types/src/pluginType.ts
@@ -1,0 +1,4 @@
+export enum pluginType {
+  Collect = "collect",
+  Store = "store"
+}
