+++ plugins/enchufa-recopiladora-bigquery/src/index.ts
@@ -2,7 +2,7 @@ import { Plugin, PluginType } from '@ingenierias-lentas/recopiladora-types';
 
 export const PluginBigquery: Plugin = {
   name: 'PluginBigquery',
-  type: 'collect',
+  type: pluginType.Collect,
   execute: () => {
     console.log('Executing Plugin Bigquery');
   }
