+++ plugins/enchufa-recopiladora-google-ads/src/index.ts
@@ -2,7 +2,7 @@ import { Plugin, PluginType } from '@ingenierias-lentas/recopiladora-types';
 
 export const PluginGoogleAds: Plugin = {
   name: 'PluginGoogleAds',
-  type: 'collect',
+  type: pluginType.Collect,
   execute: () => {
     console.log('Executing Plugin Google Ads');
   }
