+++ plugins/enchufa-recopiladora-meta-business/src/index.ts
@@ -2,7 +2,7 @@ import { Plugin, PluginType } from '@ingenierias-lentas/recopiladora-types';
 
 export const PluginMetaBusiness: Plugin = {
   name: 'PluginMetaBusiness',
-  type: 'collect',
+  type: pluginType.Collect,
   execute: () => {
     console.log('Executing Plugin Meta Business');
   }
