+++ plugins/enchufa-recopiladora-postgresql/package.json
@@ -30,7 +30,7 @@
   },
   "homepage": "https://github.com/ingenierias-lentas/enchufa-recopiladora-postgresql#readme",
   "dependencies": {
-    "@ingenierias-lentas/recopiladora-types": "^0.0.1",
+    "@ingenierias-lentas/recopiladora-types": "^0.0.3",
     "dotenv": "^16.4.5",
     "pg": "^8.12.0",
     "pg-hstore": "^2.3.4",
