+++ plugins/enchufa-recopiladora-postgresql/src/index.ts
@@ -2,7 +2,7 @@ import { Plugin, PluginType } from '@ingenierias-lentas/recopiladora-types';
 
 export const PluginPostgresql: Plugin = {
   name: 'PluginPostgresql',
-  type: 'store',
+  type: pluginType.Store,
   execute: () => {
     console.log('Executing Plugin Postgresql');
   }
